index: carousel: update css for images
Not entirely sure why, but the couple of line changes in the
carousel.css (in this commit) make the carousel images behave as I
originally wanted (i.e. the images respond to screen size changes, but
do not get too small on small screens or incorrectly move the captions
below the images).