about: start revising to show past instructors
Begin revisions to the About page to show all instructors for ENG 10
(past, present and future).

Also, to avoid overloading the page, stick with showing only instructors
(not TAs or tutors). We could show TAs and tutors, but it's a lot of
content in what should be a small page.