gallery: begin revisions to add photos from new quarters
Revise to show photos from Winter and Spring 2017.

In the future, we may want to switch to showing a collection of photos
per academic year (instead of per quarter). Or just keep a set of photos
of constant size (e.g. 20 photos) which we update semi-regularly, which
cover the main parts of ENG 10 (e.g. lecture, programming, project and
final presentations).